changes in prescription module

+++ 10.2.0/aspnet-core/src/EMRSystem.Application/Prescriptions/Dto/CreateUpdatePrescriptionDto.cs
@@ -18,7 +18,7 @@ namespace EMRSystem.Prescriptions.Dto
         public long? AppointmentId { get; set; }
         public long? DoctorId { get; set; }
         public long? PatientId { get; set; }
-        public List<int> LabTestIds { get; set; }
+        public List<long> LabTestIds { get; set; }
         public List<CreateUpdatePrescriptionItemDto> Items { get; set; }
         public List<CreateUpdateSelectedEmergencyProceduresDto> EmergencyProcedures { get; set; }
         public bool IsEmergencyPrescription { get; set; }
